big change on colours and label axis plot
