+++ src/qibo/ui/__init__.py
@@ -1,2 +1,2 @@
 from qibo.ui.mpldrawer import plot_circuit
-from qibo.ui.plot_state import plot_density_hist
+from qibo.ui.result_visualization import plot_density_hist, visualize_state
